requirements.txt
Flask==2.3.3
pygame==2.5.2
gTTS==2.5.1
SpeechRecognition==3.10.0
google-generativeai==0.4.1
requests==2.31.0
